+++ src/eip7594/fft.c
@@ -31,7 +31,7 @@
  *
  *   fr_t a;
  *   fr_from_uint64(&a, 7);
- *   for (size_t i = 0; i < 4; i++)
+ *   for (uint64_t i = 0; i < 4; i++)
  *       printf("%#018llxL,\n", a.l[i]);
  */
 static const fr_t RECOVERY_SHIFT_FACTOR = {
@@ -47,7 +47,7 @@ static const fr_t RECOVERY_SHIFT_FACTOR = {
  *   fr_t a;
  *   fr_from_uint64(&a, 7);
  *   fr_div(&a, &FR_ONE, &a);
- *   for (size_t i = 0; i < 4; i++)
+ *   for (uint64_t i = 0; i < 4; i++)
  *       printf("%#018llxL,\n", a.l[i]);
  */
 static const fr_t INV_RECOVERY_SHIFT_FACTOR = {
@@ -74,14 +74,14 @@ static const fr_t INV_RECOVERY_SHIFT_FACTOR = {
  * @param[in]   n               Length of the FFT, must be a power of two
  */
 static void fr_fft_fast(
-    fr_t *out, const fr_t *in, size_t stride, const fr_t *roots, size_t roots_stride, size_t n
+    fr_t *out, const fr_t *in, uint64_t stride, const fr_t *roots, uint64_t roots_stride, uint64_t n
 ) {
-    size_t half = n / 2;
+    uint64_t half = n / 2;
     if (half > 0) {
         fr_t y_times_root;
         fr_fft_fast(out, in, stride * 2, roots, roots_stride * 2, half);
         fr_fft_fast(out + half, in + stride, stride * 2, roots, roots_stride * 2, half);
-        for (size_t i = 0; i < half; i++) {
+        for (uint64_t i = 0; i < half; i++) {
             blst_fr_mul(&y_times_root, &out[i + half], &roots[i * roots_stride]);
             blst_fr_sub(&out[i + half], &out[i], &y_times_root);
             blst_fr_add(&out[i], &out[i], &y_times_root);
@@ -102,13 +102,13 @@ static void fr_fft_fast(
  * @remark The array lengths must be a power of two.
  * @remark Use fr_ifft() for inverse transformation.
  */
-C_KZG_RET fr_fft(fr_t *out, const fr_t *in, size_t n, const KZGSettings *s) {
+C_KZG_RET fr_fft(fr_t *out, const fr_t *in, uint64_t n, const KZGSettings *s) {
     /* Ensure the length is valid */
     if (n > FIELD_ELEMENTS_PER_EXT_BLOB || !is_power_of_two(n)) {
         return C_KZG_BADARGS;
     }
 
-    size_t roots_stride = FIELD_ELEMENTS_PER_EXT_BLOB / n;
+    uint64_t roots_stride = FIELD_ELEMENTS_PER_EXT_BLOB / n;
     fr_fft_fast(out, in, 1, s->roots_of_unity, roots_stride, n);
 
     return C_KZG_OK;
@@ -125,19 +125,19 @@ C_KZG_RET fr_fft(fr_t *out, const fr_t *in, size_t n, const KZGSettings *s) {
  * @remark The array lengths must be a power of two.
  * @remark Use fr_fft() for forward transformation.
  */
-C_KZG_RET fr_ifft(fr_t *out, const fr_t *in, size_t n, const KZGSettings *s) {
+C_KZG_RET fr_ifft(fr_t *out, const fr_t *in, uint64_t n, const KZGSettings *s) {
     /* Ensure the length is valid */
     if (n > FIELD_ELEMENTS_PER_EXT_BLOB || !is_power_of_two(n)) {
         return C_KZG_BADARGS;
     }
 
-    size_t stride = FIELD_ELEMENTS_PER_EXT_BLOB / n;
+    uint64_t stride = FIELD_ELEMENTS_PER_EXT_BLOB / n;
     fr_fft_fast(out, in, 1, s->reverse_roots_of_unity, stride, n);
 
     fr_t inv_len;
     fr_from_uint64(&inv_len, n);
     blst_fr_inverse(&inv_len, &inv_len);
-    for (size_t i = 0; i < n; i++) {
+    for (uint64_t i = 0; i < n; i++) {
         blst_fr_mul(&out[i], &out[i], &inv_len);
     }
     return C_KZG_OK;
@@ -198,7 +198,7 @@ static void g1_fft_fast(
  * @remark The array lengths must be a power of two.
  * @remark Use g1_ifft() for inverse transformation.
  */
-C_KZG_RET g1_fft(g1_t *out, const g1_t *in, size_t n, const KZGSettings *s) {
+C_KZG_RET g1_fft(g1_t *out, const g1_t *in, uint64_t n, const KZGSettings *s) {
     /* Ensure the length is valid */
     if (n > FIELD_ELEMENTS_PER_EXT_BLOB || !is_power_of_two(n)) {
         return C_KZG_BADARGS;
@@ -221,7 +221,7 @@ C_KZG_RET g1_fft(g1_t *out, const g1_t *in, size_t n, const KZGSettings *s) {
  * @remark The array lengths must be a power of two.
  * @remark Use g1_fft() for forward transformation.
  */
-C_KZG_RET g1_ifft(g1_t *out, const g1_t *in, size_t n, const KZGSettings *s) {
+C_KZG_RET g1_ifft(g1_t *out, const g1_t *in, uint64_t n, const KZGSettings *s) {
     /* Ensure the length is valid */
     if (n > FIELD_ELEMENTS_PER_EXT_BLOB || !is_power_of_two(n)) {
         return C_KZG_BADARGS;
@@ -254,7 +254,7 @@ C_KZG_RET g1_ifft(g1_t *out, const g1_t *in, size_t n, const KZGSettings *s) {
  *
  * @remark The coset shift factor is RECOVERY_SHIFT_FACTOR.
  */
-C_KZG_RET coset_fft(fr_t *out, const fr_t *in, size_t n, const KZGSettings *s) {
+C_KZG_RET coset_fft(fr_t *out, const fr_t *in, uint64_t n, const KZGSettings *s) {
     C_KZG_RET ret;
     fr_t *in_shifted = NULL;
 
@@ -285,7 +285,7 @@ out:
  * @remark The coset shift factor is RECOVERY_SHIFT_FACTOR. In this function we use its inverse to
  * implement the IFFT.
  */
-C_KZG_RET coset_ifft(fr_t *out, const fr_t *in, size_t n, const KZGSettings *s) {
+C_KZG_RET coset_ifft(fr_t *out, const fr_t *in, uint64_t n, const KZGSettings *s) {
     C_KZG_RET ret;
 
     ret = fr_ifft(out, in, n, s);
